success with mismatch outputs

+++ code/lib3.py
@@ -1,8 +1,8 @@
 def getPrice(finance):
-  return int(finance.get("price", 0))
+  return float(finance.get("price", 0))
 
 def getRevenue(finance):
-  return int(finance.get("revenue", 0))
+  return float(finance.get("revenue", 0))
 
 def getProfit(finance):
   return getPrice(finance) - getRevenue(finance)
